test(db): fix plan rpc permission and bigint assertion

+++ supabase/tests/08_plan_functions.sql
@@ -50,7 +50,7 @@ SELECT
             SELECT count(*)
             FROM get_current_plan_max_org('22dbad8a-b885-4309-9b3b-a09f8460fb6d')
         ),
-        0,
+        0::bigint,
         'get_current_plan_max_org test - anonymous call cannot read org plan limits'
     );
